+++ tool_sideview/sideview_view.py
@@ -698,9 +698,8 @@ class SideViewDockWidget(QDockWidget):
             self.iface.mapCanvas().unsetMapTool(self.route_tool)
             self.select_sideview_button.setChecked(False)
 
-            # Route tool is unset, clean up virtual tree and current route
+            # Route tool is unset, clean up virtual tree
             self.route.reset()
-            self.map_visualisation.reset()
             self.select_sideview_button.setText("Choose side view trajectory")
 
     def toggle_route_tool(self):