+++ CHANGES.rst
@@ -15,6 +15,7 @@
 - Graph: added batch delete option (#884)
 - Graph: features with same id, but different layer should have different color (#895)
 - Statistics: fix python error with "Bed level gradient" (#876)
+- Statistics: assume has_interflow for older models (#889)
 - Visualization: Remove legend subgroup hidden feature (#769)